Revert "Add handler for git upload-archive to Diffusion"
The handler in this commit doesn't work correctly: git-upload-archive
doesn't do full command line option parsing (argv[1] is simply treated
as the repository path), so passing "-- <repo_path>" as arguments to
git-upload-archive in the call to sudoCommandAsDaemonUser() results in

    0008ACK
    00000026usage: git upload-archive <repo>

being printed to stdout. The solution is simply to pass "<repo_path>" as
the sole argument.

This reverts commit 5ecea584edfdacaa12b0f4bdf52046aa2a8b9557.