try squashmerge first then rebase (#74)
What:
Previously we had patched the landing workflow to first try to rebase the revision with the commit we are trying to land onto. This is generally a branch. Then if we fail to rebase we revert back to the original simpler workflow of squash merging the head of the revision onto the branch. Instead this PR swaps the order of attempt so we first try to squash merge then if that fails we rebase.

Why:
We discovered that some developers amend merge commits to include changes unrelated to merge conflicts. Git treats all changes part of a merge commit as being merge conflict resolutions so when you rebase it will drop all changes in a merge commit. This meant some changes in peoples revisions were not included in the landed commit. I decided it was safer to try the merge first since this has been tested for a lot longer and people are used to the behaviour. We still attempt the rebase if there is a stacked revision though since this improves the workflow of people who stack revisions. I have also patched arc so that it doesn't amend merge commits by default so the instances of amended merge commits will decrease considerably over time making the rebase much safer.
---------

Co-authored-by: rgodden <rgodden@thoughtmachine.net>