remove diff onto herald field and add revision onto herald field (#81)
The Diff herald rules are evaluated on new diffs uploaded as part of the conduit differential.creatediff which doesn't have the capability to send onto branch information. This means we can't use herald rules on onto branches since we don't have the onto branch yet. Revisions do have the onto branch available to them so create a herald field for revisions instead.

---------

Co-authored-by: rgodden <rgodden@thoughtmachine.net>